An initial commit of a working program
This is the initial commit with the firmware compiling as it should. Have
not yet tried it on actual hardware. The current implementation is quite
naive, with just a set PWM cycle for each ramp, and just on/off for holding
a temperature. Will need some refactoring to become managable soon too.