Move to c++11 and print elapsed time
Fixed makefile to use c11 and gnu++11 as standards. Also added the flag
for optimization for easy experimentation.

Fixed some warnings about signed with unsigned comparisons (profile is
now only unsigned values).

Move the temperature print right, and print elapsed time (both total and
for the specific stage) to the left of it.