Update cmake-files, and fix size problem
Updated the arduino-cmake project to a more modern version, with at
least partial support for Arduino SDK 1.6.12, and that also remove a
bunch of warnings and some other bugs.

Fixed the flag settings in CMakeLists.txt so that the program size
gets more reasonable. It's now down to half the size, and is the same
as it was with Arduino-Makefile.